Refine LiterateCheck workflow to target README.md
Updated the workflow to check for changes specifically in README.md instead of all modified files. This is needed in recent versions of Julia v1.12 since Pkg now adds `[sources]` to Project.toml files automatically when adding based on URLs or relative paths, and we need to ignore those changes.